ansible/hg-ssh-server: change `scm_allow_direct_push` gid to 692 (Bug 1515119) r=glob
`2500` was selected as the `gid` for the new scm group
completely at random. Turns out, that value is important
and will block the deploy during creation since it doesn't
match the value on the hgssh servers.

This commit changes the value to 692, which should fix
the deploy blocker. This is untested but is harmless in
the event it doesn't work as we want it to.