hghooks/lando_required: refer to most recent topological commit as "head" (Bug 1515119) r=glob
Using the term "top" is confusing, since it could refer to either
the commit at the "base" of the stack, or the "tip" of the stack.
Many VCS' use "head" to refer to a commit with no children,
so this commit changes the Lando required hook to use the less
ambiguous term.

Differential Revision: https://phabricator.services.mozilla.com/D39072

--HG--
extra : moz-landing-system : lando