hghooks/lando_required: disable demandimporter when importing Sentry SDK (Bug 1515119) r=glob
When testing the Lando required hook, I noticed that I was
able to send notifications from the pash virtualenv on
hgssh1 using the Python REPL, but could not get the message
through to Sentry when performing a push. After some testing
I discovered the Sentry SDK does not like Mercurial's
demandimporter. This commit disables the importer during the
Sentry SDK import.

We also move the import into the logging function. This
is both for performance, and to avoid an `ImportError`
in the `unifyrepo` process, which does not have the
Sentry SDK installed and fails when attempting to import
the SDK without demandimport.

Differential Revision: https://phabricator.services.mozilla.com/D39071

--HG--
extra : moz-landing-system : lando