hghooks/lando_required: add more context to Sentry notifications (Bug 1515119) r=glob
This commit adds more context to the Sentry event messages.

The core message that is displayed in Slack is extended to add
the "justification" for using a manual push, and some unnecesary
cruft is removed. The two unique message formats (one for each
acceptable scm level) are consolidated into a single message
with the scm level as a parameter. As part of this change,
`changeset_has_justification` is renamed and changed from a
boolean check into a function which returns a value or `None`.

The Sentry APIs are used more extensively, adding user metadata
as well as some "tags" and extra pieces of information. This
will make organizing events by different parameters easier in
Sentry. For example, determining which repo is used most
frequently for direct pushes, which user is pushing directly
most often, etc.

Since users with `scm_allow_allow_direct_push` do not need to
provide justification for their pushes, a default string is
inserted in place of the justification.

Differential Revision: https://phabricator.services.mozilla.com/D39209

--HG--
extra : moz-landing-system : lando