mozext: unify into a single test and improve coverage (Bug 1569734) r=smacleod
The test coverage for `mozext` was quite lacking, and contributed
to a lot of the bit rot in the last few months. In this commit I
create a single unified test that covers most of the added template
keywords, revset predicates and commands.

Since we rely on the `pushlog` extension to enable several features,
we pin the Mercurial used in `hg serve` to a known-compatible version.

Changing this test and pinning the version of `hg` used to serve the
pushlog allow us to remove the test from the flaky test list.

Differential Revision: https://phabricator.services.mozilla.com/D39955

--HG--
rename : hgext/mozext/tests/test-pushlog.t => hgext/mozext/tests/test-mozext.t
extra : moz-landing-system : lando