mozext: add `hg oops` command from `qbackout` (Bug 1565933) r=smacleod
In changeset b3a88dcabb6cccf0 we removed the `mq` related extensions
still checked in to version-control-tools. It turns out that the
`qbackout` extension has a command `hg oops` which is not dependent
on `mq` and is used heavily in the sheriffing workflow. This commit
moves the extension into `mozext` to keep it alive for the sheriffs.

Minimal changes have been made from the version in `qbackout`, and
the test is completely unmodified.

Differential Revision: https://phabricator.services.mozilla.com/D39953

--HG--
extra : moz-landing-system : lando