mozext: remove `hg mybookmarks` (Bug 1569734) r=smacleod
Usually the bookmarks in a repo are either the users own, or
references to the various tree heads (central, inbound, etc).
If the user is using `firefoxtree`, they won't see the fxheads
bookmarks in `hg bookmarks` or `hg show bookmarks` output. So
this solution is older and can be removed.

Differential Revision: https://phabricator.services.mozilla.com/D39957

--HG--
extra : moz-landing-system : lando