hgtemplates: escape filenames in pushlog ATOM feed (Bug 1575686) r=glob
A few of the Thunderbird devs noticed the ATOM pushlog feed was busted
recently. Upon further inspection it appears we have not been escaping
filenames in the ATOM feed since it's inception (~10 years ago!).

This commit passes the filename through the `escape` template filter.
We also add a new push/changeset to the test setup, with a filename that
includes an `&` character, to confirm the results are escaped in
`test-hgweb-atom.t`. This causes a few other tests to change as well,
since the new push will be included in output of those tests.

Differential Revision: https://phabricator.services.mozilla.com/D43082

--HG--
extra : moz-landing-system : lando