ansible/monitoring-agent: install Telegraf from InfluxData yum repo (Bug 1576212) r=smacleod
Telegraf is busted on hgweb* in MDC1 due to the mozilla-infra yum repo
installing a bad version of Telegraf over the pinned version in v-c-t.
The recommended installation method is using the InfluxDB yum repo,
so this commit switches the `monitoring-agent` Ansible role to add the
new repo and install Telegraf through it.

Differential Revision: https://phabricator.services.mozilla.com/D44889

--HG--
extra : moz-landing-system : lando