Bug 1584954 - hgext: fix a few issues with mach formatting hook extensions r=smacleod
A user reported an issue with the mach formatting hooks on IRC,
reproducible with the following command:

  $ hg init && echo foo > bar && hg commit -A -m initial
  nothing changed
  [1]

The message "nothing changed" and a return code of 0 are not
expected. Running `hg log` shows the commit still exists despite
the message. The cause of this issue was having a call to the
original wrapped function in the `finally` block. If we hit the
"no files changed" or "is not a Firefox repo" conditional in the
`try` block, this would cause a double call to `orig`, with the
second call being the source of the bad error message and return
code.

While I was fixing this bug, I decided it would be easier to move
the "is this a Firefox repo" logic into the `reposetup` stage of
the extension setup, which short-circuits the function wrapping on
irrelevant repos. We also move the check for the `MOZPHAB`
environment variable to the same location, and add the check to
`js-format` which was overlooked when this was added.

Finally we add test coverage for the reported case and also add
a simple test of the `MOZPHAB` environment variable behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D47150

--HG--
extra : moz-landing-system : lando