serverlog: limit application layer logging to relevant hosts r=smacleod
We log SSH connections on the HTTP server and HTTP connections
on the SSH server. This is due to the config flags defaulting
to `True`, and not being set in the respective system hgrc files
on each server type. This commit changes the defaults to False
and selectively enables the necessary logging per-server.

Differential Revision: https://phabricator.services.mozilla.com/D48226

--HG--
extra : moz-landing-system : lando