serverlog: replace call stack inspection with a flag and more function wraps r=smacleod
Previously we would walk the stack and check for the presence of a specific function
name, only registering a completed command in the absence of the function (`_getpayload`).
In this commit we switch to wrapping `_getpayload` and setting a flag on the `_serverlog`
mapping, registering a completed command when the flag is not set.

While we're here, we add byte string prefixes to some config knobs.

Differential Revision: https://phabricator.services.mozilla.com/D48227

--HG--
extra : moz-landing-system : lando