scripts: add `gcp` option to `scrape-manifest-ip-ranges.py` (Bug 1585133) r=smacleod
This commit adds a new subcommand to `scrape-manifest-ip-ranges.py`,
which scrapes Google's DNS records to gather information about it's
public IP address blocks. The process implemented in this commit is
outlined in Google's cloud support docs. [1]

To summarize, we use the `dnspython` DNS toolkit to first run a
query for `_cloud-netblocks.googleusercontent.com`. This query
will return a list of domains, each returning a set of IP blocks
for Google Cloud Platform services. The resulting blocks are then
saved to a file on disk. An example output looks like:

  ip4:35.199.0.0/17
  ip4:35.199.128.0/18
  ip4:35.235.216.0/21
  ip6:2600:1900::/35
  ip4:35.190.224.0/20

[1] https://cloud.google.com/compute/docs/faq#find_ip_range

Differential Revision: https://phabricator.services.mozilla.com/D49295

--HG--
extra : moz-landing-system : lando