terraform: create resources to store Mercurial clonebundles in GCP (Bug 1585133) r=smacleod
This commit adds Terraform configs for a GCS bucket and service account
required to publish Mercurial clonebundles to GCP. The service account
represents the hgssh master server process which generates the bundle
and uploads to GCP, with the corresponding key being used for credentials.
The bucket is created with a retention policy and lifecycle policy of
7 days. The retention policy holds data as undeletable for a minimum of
7 days and the lifecycle policy deletes the data after it is 7 days
past expiration time.

Differential Revision: https://phabricator.services.mozilla.com/D49512

--HG--
extra : moz-landing-system : lando