hgmo: prioritize stream clone bundles when cloning from GCP (Bug 1585133) r=smacleod
This commit teaches the hgmo extension to prioritize stream clone bundles
when responding to clone requests from IP addresses in GCP. To do so we
make the `filter_manifest_for_aws_region` more generic to account for the
new GCP regions. We add a new config option which points to a path on disk
where the previously added GCP IP scraper will dump a file containing IP
addresses for known GCP blocks. This file is mocked out by adding an
example file to the `docker-hg-web` Ansible role.

Differential Revision: https://phabricator.services.mozilla.com/D49514

--HG--
extra : moz-landing-system : lando