mozautomation: remove `Build` namedtuple in favor of raw dict and remove `artifacts_url` parameter (Bug 1563138) r=glob
To make this class work with Mercurial under Python 3, we would have
to override the `_asdict()` method to return a dict with byte-string
keys, or encode/decode whenever we need to work with values in the
dict. Instead, this commit removes the `Build` namedtuple class
and converts all consumers to use the dict, which now uses byte
strings by default. We also remove the unnecessary internal parameter
`artifacts_url`, converting to `archive_url` everywhere as is exposed
in the various APIs.

Differential Revision: https://phabricator.services.mozilla.com/D38701

--HG--
extra : moz-landing-system : lando