hgext/firefoxtree: use `util.safehasattr` to determine object type (Bug 1541166) r=glob
Previously we were using `isinstance(obj, str)`. This solution isn't
compatible with Python 2 and 3. A simpler solution is to check if
the object has the `local` attribute. If it does, it is a repo
instance, and we want to call the function. Otherwise, we create
the local repo ourselves.

While we're here, we delete the comment regarding a standalone function
that we should migrate to, as it doesn't appear to exist and similar
code exists within modern versions of Mercurial.

Differential Revision: https://phabricator.services.mozilla.com/D37827

--HG--
extra : moz-landing-system : lando