hgext/firefoxtree: convert `template` key to byte string before passing to template formatter (Bug 1541166) r=glob
The `fxheads` command allows passing the `-T`/`--template` argument.
This command line flag is converted to a Python keyword argument when
passed to the function that runs the specified command. As a result,
the key in the `opts` dict (which is essentially just a renamed
`**kwargs` call) is a unicode string under Python 3, while the template
formatter object is expecting a dict with a byte string `template`
argument. In this commit we manually pass a dict with the correct
byte string key.

Differential Revision: https://phabricator.services.mozilla.com/D37828

--HG--
extra : moz-landing-system : lando