hgext/firefoxtree: pass `rev` argument in a list (Bug 1541166) r=glob
This commit fixes what is actually a longstanding bug in `firefoxtree`
that by pure coincidence has been providing the intended behaviour.

Essentially the value of the `rev` option passed to most Mercurial
commands in the Python function definiton is supposed to be a list
of revsets (ie one entry for each `-r <revset>` used on the command
line). In `firefoxtree`'s wrapping of `hg outgoing`, we override the
value of `rev` and always default to using `.` as our revset, except
we don't pass our overridden value in a list.

In many code paths that use the `rev` option (revset parsing code,
primarily) the elements of the `rev` list are iterated over, and
each is evaluated to be converted into symbols for the revset language.
Since we are passing a single string instead of a list of strings to
iterate over, we were accidentally iterating over the characters in
our single revset.

On Python 2, when iterating over elements of a `str`, you get the
string representation of those elements. For example, iterating over
`'hello'` would give the sequence `['h', 'e', 'l', 'l', 'o']`.
For our revset of `'.'`, by pure luck, our iteration would
end up giving the correct sequence, `['.']` (since we only have a
single element in the string).

On Python 3, we need to use `bytes`, as Mercurial assumes all strings
are represented in 8 bits. When iterating (or doing any accesses of
elements) over a byte string, instead of returning the string
representation as Python 2 does, we get the integer representation
of those bytes[0]. In the case of `b'.'`, this means our iteration
sequence looks like `[46]`. Later down the line, the revset parser
sees `46` and assumes this is a revision number, which it later uses
to search the index, causing an `IndexError`.

The correct way to pass our revset override is to wrap it in a list.
So this commit fixes the bad option, and brings `firefoxtree` up
to Python 3 compatibility.

[0] https://docs.python.org/3.7/library/stdtypes.html?highlight=bytes#bytes.hex,
    see directly below the definition of `hex`

Differential Revision: https://phabricator.services.mozilla.com/D37830

--HG--
extra : moz-landing-system : lando