hgext/pushlog: wraps values from `sqlite3` as compatible byte strings (Bug 1541166) r=glob
This commit adds parameters to the `sqlite3` connection object
to force the returned objects into the correct values. The
`detect_types` keyword argument is set to `PARSE_DECLTYPES`,
which means that all returned values are returned as the type
in their schema defintion (ie something defined as "integer
primary key id" will be passed through the "integer" converter
before returning to the caller). The `text_factory` function,
which takes a single bytestring parameter and returns the
resulting object, is set to `pycompat.bytestr` to ensure all
`TEXT` values returned by `sqlite` are compatible with both
Python 2 and 3.

`sqlite3` also requires values used in queries be the `str` type
used by the Python version in question. We wrap all values passed
as parameter substitutions in `pycompat.sysstr`.

Differential Revision: https://phabricator.services.mozilla.com/D38706

--HG--
extra : moz-landing-system : lando