hgext/hgmo: create a shallow copy of `entry` for iteration (Bug 1541166) r=glob
In the `hgmo` extension we iterate over a dict representing
a changeset "entry", mutating the state of the dict as we
do so. This includes some `del` statements and changes to
various values in the dict.

On Python 3, mutating a dict during iteration is not allowed
and throws a `RuntimeError`. This commit works around that
restriction by creating a copy of the dict and iterating over
it, while mutating the original object.

Differential Revision: https://phabricator.services.mozilla.com/D38711

--HG--
extra : moz-landing-system : lando