hgext/overlay: add byte string prefixes everywhere (Bug 1563138) r=glob
This gets us to all tests passing with one exception.
In that case the test output is correct, but the `hghave`
for Py3 seems to be failing to recognise the test is
running against Python 3. It appears the Py3 check
compares against the value of `sys.version_info`,
and I believe in this case is grabbing the version
from the interpreter of the test runner instead of
the Python 3 `hg`.

Differential Revision: https://phabricator.services.mozilla.com/D38716

--HG--
extra : moz-landing-system : lando