py3: use `pycompat.{strkwargs,byteskwargs}` to ensure arguments are correct type (Bug 1563138) r=glob
Changing most strings to byte-strings has the unfortunate side effect
of making those strings unusable as keyword arguments via dictionary
unpacking. The `pycompat` module has two helper functions that can
convert the top-level keys in a dict to unicode strings (for passing
to Python functions) or to byte strings (for passing to various
Mercurial functions, or passing the whole dict as an argument). This
commit wraps occurrances of mismatched keys to ensure compatibility
with Python 3.

In the case of the wrapped `webutil.changesetentry`, the function is
only called once in upstream Mercurial code, and it's return value
is immediately unpacked to be passed as kwargs to another function.
Since we want to pass the value of that function to `addmetadata`,
we need to first convert the keys to bytestrings, then re-convert
the keys to unicode strings before passing back to Mercurial. The
wrapper for `webutil.changelistentry` is called from within Mercurial
code and is not passed as keyword arguments, so we don't wrap those
values before passing to `addmetadata`.

Differential Revision: https://phabricator.services.mozilla.com/D38712

--HG--
extra : moz-landing-system : lando