mozautomation: use set of build IDs to filter previous builds (Bug 1591955)
Previously we created a set of `Build` namedtuples to check which
values builds we have already seen and avoid duplicating entries.
Since we switched to a dict, we can no longer use that technique.

This commit changes the set we use to track duplicates to instead
be a set of build IDs, which are able to be turned into a set.

--HG--
extra : amend_source : 313bc4626ffd020eed9397707d797eedd6ab9ce4