testing: add a Py3 test ignore list and teach `run-tests` to respect it by default (Bug 1541166) r=glob
Similar to our "flaky" tests list, we introduce a list of
tests which fail under Python 3 and enforce it's use when
running tests with a Python 3 `hg`. To do so we add a flag
`--run-py3-failing` which performs the same function as the
flaky test list (add the blacklist unless the flag is present)
except it also asserts the Python version used in the `hg`
binary is Python 3.

To get this information, we expand `get_hg_version` to return
the Python version as well as the Mercurial version for a
given `hg`. We also allow `subprocess.CalledProcessError` to
be called, since sometimes the error code can be `1` despite
the correct JSON output being emitted. We will stil fail when
attempting to decode the JSON if some fatal event has actually
occurred. Making this change requires us to move a check to an
earlier point in the script, which should have no effect.

Differential Revision: https://phabricator.services.mozilla.com/D50325

--HG--
extra : moz-landing-system : lando