vcsreplicator: convert `hg debugbase85obsmarkers` to use the templater (Bug 1541166) r=glob
Using `json.loads` to turn the data into JSON would require
type-checking all fields to ensure they are the `str` type
before attempting to serialize. We also switch to using the
formatter's `formatdate` function, which enforces the correct
Mercurial date format in it's output. As a result of this a
previous assumption that the first value in the date tuple is
a float is invalidated, so that value is wrapped in `float`
call.

Differential Revision: https://phabricator.services.mozilla.com/D50330

--HG--
extra : moz-landing-system : lando