vcsreplicator: upgrade `kafka-python` to 1.4.6 (Bug 1557783) r=glob
`kafka-python` version 1.4.2 has a bug that causes an `ImportError`
when loading `kafka.producer.simple` due to a variable named `async`
which is a keyword under Py3. This problem is fixed in version 1.4.6,
so this commit upgrades to the new version in all virtualenvs in
which we install that package.

Differential Revision: https://phabricator.services.mozilla.com/D50331

--HG--
extra : moz-landing-system : lando