hgext/hgmo: use `functools.cmp_to_key` to work around Py3 `sorted` differences (Bug 1541166) r=glob
Python 3 changed the `sorted` function to take a `key` function
(a function which can be used as a sorting key for the items in
the sequence) instead of a `cmp` function (a function which takes
two arbitrary items and orders them). This commit papers over the
differences using `functools.cmp_to_key`. We can fully rewrite this
as a key function when we drop Python 2 support.

Differential Revision: https://phabricator.services.mozilla.com/D50334

--HG--
extra : moz-landing-system : lando