upgrade-python-packages: support passing a custom `pip-compile` to the script r=glob
We have a custom wrapping script for `pip-compile` that we use
to upgrade dependencies in the repo. This script calls `pip-compile`
natively from the command line, assuming it is somewhere on the users
`PATH`. The script also requires Python 3 to run as it takes
advantage of some new features and libraries. As a result, the way
I have been using the script is by creating a Py3 virtualenv with
`pip-tools` installed and running it from there.

After some investigation into the recent bundle virtualenv breakage,
I realized this assumption about usage of the script is flawed.
`pip-tools` method of supporting different Python interpreters is
to run it from a virtualenv of the specific Python version you want
to support. This means that to generate Python 2 requirements you
should install `pip-tools` a Python 2 virtualenv, and either put
that virtualenv on your path or run `pythonX.Y -m piptools compile`.

This commit adds support for passing a `pip-compile` executable via
a command line argument, defaulting to the local `pip-compile` if none
is specified. This allows us to use `pip-compile` on Python 2 (or really
any other Python version that is not the same as the virtualenv running
the upgrade script) without much modification to the underlying script.

See here for more:
    https://github.com/jazzband/pip-tools#example-usage-for-pip-compile

Differential Revision: https://phabricator.services.mozilla.com/D50980

--HG--
extra : moz-landing-system : lando