hgmo: move instance data JSON path to config r=glob
This value has been hard-coded into the extension since its
inception. Since it is a path to a config file which is subject
to change, I believe it makes more sense if we specify it as a
config knob. This commit moves the path into the hgrc config for
the hgweb machines. We only add the path if the worker has
replication filters, which at the moment doubles as an indicator
that the instance being deployed to in Ansible is a cloud mirror.

Differential Revision: https://phabricator.services.mozilla.com/D51179

--HG--
extra : moz-landing-system : lando