hgext: mark extensions as compatible with Mercurial 5.1 (Bug 1569741) r=glob
No code changes required for passing tests here. This includes running
the tests using Mercurial 5.1 in the Docker image virtualenvs.

The only change in tests is the removal of extension setup timings in
blackbox log. This feature is moved behind a debug flag in 5.1+ (which
is probably more correct behaviour). These lines will be removed when
we run the actual upgrade on the server.

Differential Revision: https://phabricator.services.mozilla.com/D51200

--HG--
extra : moz-landing-system : lando