ansible/monitoring-agent: add InfluxDB metric for hg bundle generate `lastrun` file age r=glob
This commit teaches Telegraf to monitor the bundle generation `lastrun`
file using the `filestat` input. This will send a ping to InfluxDB every
day with the age of the file, it's size and a boolean indicating it's
existence. We will use these metrics to create a TICKscript that alerts
when the bundle generate script fails to complete successfully.

Differential Revision: https://phabricator.services.mozilla.com/D51623

--HG--
extra : moz-landing-system : lando