testing: install a Py3 interpreter and Mercurial executables in test environment (Bug 1541166) r=glob
This commit teaches `create-test-environment` to install a Python 3
interpreter in the `venv/` directory created by the script. We
introduce new `test-requirements-3.{in,txt}` files which are
installed into the interpreter after creation. These new files are
derived from `test-requirements.{in,txt}` and include all packages
necessary to run tests for v-c-t. The command assumes the user has
`python3` aliased to a Python 3 interpreter in their shell.

The main use case for this change is to enabled running tests against
Python 3 Mercurial installations, like so:

  $ ./run-tests hgserver/ --with-hg=venv/py3/mercurials/5.2rc0/bin/hg

Support for `--all-hg-versions` will be added in a later commit.
Targetting specific versions of Python can be achieved by aliasing
`python3` to the desired interpreter in the users shell.

Differential Revision: https://phabricator.services.mozilla.com/D51735

--HG--
extra : moz-landing-system : lando