third_party: update packages and vendor `six` (Bug 1593379) r=glob
Upgrading `configobj` in changeset f696566419d6 caused a regression
as we failed to also vendor `six`, which was added as a dependency.
This new dependency was overlooked in our test environment, as we
install `six` via a transitive dependency from another package. In
the case where the user doesn't have a virtualenv installed, we will
want to have `six` vendored to allow `configwizard` to continue
working.

This commit adds `six` as a dependency and upgrades our third party
packages via the `upgrade-python-packages` script and the
`revendor-python` script. We also add `six` to the list of packages
added to the interpreter via `$PATH` manipulation.

Differential Revision: https://phabricator.services.mozilla.com/D52023

--HG--
extra : moz-landing-system : lando