firefoxreleases: use build anchor to filter builds (Bug 1594758) r=glob
In fixing bug 1591955 we changed the `previous_builds` set to be a
set of `build_id` strings. Previously this set was composed of
`Build` namedtuples, and the switch to using a dict caused failures
since a dict cannot be hashed for inclusion in a set. It turns out
the `build_id` field is not a unique identifier for the build as
I had believed, and we are now seeing missing builds in the output
of the `firefoxreleases` webcommand.

This commit moves the `build_anchor` function from `firefoxreleases`
into `mozautomation.releasedb`, and changes the `previous_builds`
set to be built using instances of these strings. The string is a
concatenation of the short revision, release channel, platform and
build ID and should be a unique identifier.

I believe that landing and deploying this change will cause the next
iteration of the release scraper to add the missing builds to the
database.

Depends on D52270

Differential Revision: https://phabricator.services.mozilla.com/D52271

--HG--
extra : moz-landing-system : lando