ansible: remove Py2 venv if installing Py3 venv in same directory (Bug 1541166) r=glob
Upgrading to Python 3 will involve re-installing many of the various
virtual environments on hgmo and upgrading dependencies for Python 3.
Rather than re-name all the environments, this commit teaches the
virtualenv3 Ansible task to check for the presence of a Python 2
virtual environment and delete it if present, writing the new
Python 3 virtual environment in it's place.

Differential Revision: https://phabricator.services.mozilla.com/D52202

--HG--
extra : moz-landing-system : lando