hgmo: set `web.allow-archive` to empty (Bug 1596135) r=glob
As described in the bug, this command is causing load problems on
the public hgweb interface. Since we have `xz` archives available
for download elsewhere, and any archive can still be built from a
clone of the repo using `hg archive`, this commit disables the
webcommand to archive the repo server-side. We do this by setting
the `web.allow-archive` config to an empty value (ie no archives
allowed). This disables the webcommand on the server side and also
removes the various links on the web interfaces pointing at URLs
for the webcommand.

Differential Revision: https://phabricator.services.mozilla.com/D52878

--HG--
extra : moz-landing-system : lando