testing: add docker-compose env for running tests (Bug 1598958). r=sheehan
This introduces a docker-compose.yml and a test-runner
docker image for executing the tests. This image is built
with everything needed to run the non-docker tests.

when run through docker-compose the container makes use of
the host docker daemon to build and run the full docker
tests. Currently the container also requires using host
networking so that it can access the containers through
localhost. This means only a linux host is supported for
running these tests, as docker for windows and docker for
mac os do not support this feature.

Eventually as we re-implement the docker test containers
using docker files, and switch the docker-compose networking
we'll be able to remove the requirement on host networking.

Differential Revision: https://phabricator.services.mozilla.com/D54423

--HG--
extra : moz-landing-system : lando