ansible/hg-ssh: create bundles for autoland (Bug 1598215) r=smacleod
In automation we should be cloning `mozilla-unified` and pulling from
`autoland`, so this change is mostly for developers. In the event we
did do a full clone of autoland in automation, it would be better to
have bundles available, so this change is a win-win from both viewpoints.

Differential Revision: https://phabricator.services.mozilla.com/D54347

--HG--
extra : moz-landing-system : lando