robustcheckout: wrap `win32.unlink` instead of `util.unlink` during purge on Windows (Bug 1597746) r=smacleod
Rather than wrap the imported `util.unlink` in the `purge` extension,
which no longer exists as of Mercurial 4.8, we should wrap the unlink
function in `mercurial.win32` directly and call our special unlinker
when the thrown `OSError` has the `winerror` attribute. We should also
be catching `OSError` instead of `WindowsError` since the latter has
been removed in Python 3. This commit makes these changes, and marks
the new version of robustcheckout as requiring Mercurial 4.8.

Differential Revision: https://phabricator.services.mozilla.com/D54209

--HG--
extra : moz-landing-system : lando