ansible/hg-ssh: make `check_kafka_hg_heartbeat` Python 3 compatible (Bug 1541166) r=smacleod
The `check_kafka_hg_heartbeat` Nagios check has a few encoding
issues in Python 3, since the values returned by `subprocess.check_output`
are byte-strings but the rest of the script assumes regular Python
`str`'s. We also attempt to directly print an exception before
converting to a `str`, so this commit adds that conversion as well.

Differential Revision: https://phabricator.services.mozilla.com/D56210

--HG--
extra : moz-landing-system : lando