testing: fix some encoding issues in test code r=smacleod
These lines of test code run into encoding errors (ie "must be str,
not bytes"). When running under Python 3. For the time being we
won't be converting the `venv_testing` environment to Python 3,
but making these changes gives us compatibility with both versions,
so we might as well take them now.

In newer versions of `moto_server`, AWS regions are actually partitioned
as separate bucket/topic resource locations despite using the same
endpoint to create the bucket/topic. This commit changes the region we
create these resources in to `us-west-2` to align with testing code
in most other places in the codebase.

Differential Revision: https://phabricator.services.mozilla.com/D56125

--HG--
extra : moz-landing-system : lando