robustcheckout: only apply `unlink` wrapper on Windows (Bug 1597746) r=mhentges
The previous patches for this bug (changeset c2c9f8f33757) broke running
`hg purge` with `robustcheckout` on non-Windows platforms. This commit
changes the wrapper to only be applied on Windows, where it is required.

We also add a smoketest call to `hg purge` in `test-purge.t` to ensure
we can still run this command with `robustcheckout` enabled.

Differential Revision: https://phabricator.services.mozilla.com/D56410

--HG--
extra : moz-landing-system : lando