evolution: use new obsolescence config flags r=mhentges
Previously the `obsolescence` feature was hidden behind a single
`experimental.evolution` flag, with the value of that flag being a
list of comma-separated values indicating which obsolescence features
to enable (for example, `experimental.evolution=createmarkers,exchange`).
If all evolution related features were desired, passing `all` as the
value would fully enable evolution.

The config has been changed to a single `experimental.evolution` flag
which can be set as a boolean (ie `True`), or a set of piecewise boolean
flags that can be used to enable features individually (for example
`experimental.evolution.createmarkers=true`).

This commit changes all instances of this flag to the new format. We also
modify the `obsolescencehacks` extension to turn on obsolescence using
the new flag.

Differential Revision: https://phabricator.services.mozilla.com/D56630

--HG--
extra : moz-landing-system : lando