configwizard: change `hgversion` check default to "upgrade" instead of "no-upgrade" (Bug 1602204) r=sylvestre
Currently the default options in configwizard all move the user towards
the preferred configuration state (for example, enabling extensions in
their config), with the exception of upgrading their Mercurial version,
which defaults to "no".

This commit tweaks the wording of the prompt for the `hgversion` check
to exit when the user is on an unsupported version of Mercurial and
should upgrade.

Differential Revision: https://phabricator.services.mozilla.com/D56390

--HG--
extra : moz-landing-system : lando