hooks: adding push hook for security bugs (Bug 1420510) r=sheehan
This revision implements a new hook, "check_bug_references" that
checks bug IDs referenced in all commit messages for the presence of any invalid
bugs or security bugs, and rejects such pushes and displays a message to the
user.

The check goes through each commit that is in draft phase and parses any bug IDs
to be checked at once against the BMO REST API's search endpoint. If any bugs
are filtered out, the check then checks a single bug ID of the invalid bugs
against the API to determine the reason why the bug is invalid.

If the BMO REST API service is down, checks can not be performed, and as such
commits will be rejected by default. Users can try pushing their changes at a
later time in that case, or can include a flag in the commit message to skip
these checks.

As part of this revision, a minimal wrapper was created that abstracts
interactions with the BMO REST API. As a future change, it would be good to
move this wrapper (`BMOAPIClient`) elsewhere into its own module.

The url-interceptor.py module will also need to be deprecated in a future
revision, as the `advanced_url_interceptor` module was created to replace it.
This new module allows contributors to mock HTTP responses that are triggered
via the `mercurial.urllibcompat` module using a simple JSON file.

A new configuration key, "check_bug_references_repos"  was added under the
"mozilla" section in hgrc to list the repositories that this check should apply
to. Initially this will be set to "try" repo on hgssh.

Differential Revision: https://phabricator.services.mozilla.com/D55290

--HG--
extra : moz-landing-system : lando