ansible/hg-web: remove Mozilla IP scraper r=mhentges,smacleod
The Python version of `bloxtool` we're using to query infoblox
doesn't support Python 3. Since Infoblox hasn't been shown to
return the correct IP blocks anyways, I think we should just
remove the scraper and advance with the Py3 upgrade.

The code to prefer stream-clone bundles in the `hgmo` extension
has been left as-is for now. We can finish pulling that out, or
convert to the Rust version of Infoblox sometime next year.

Differential Revision: https://phabricator.services.mozilla.com/D56964

--HG--
extra : moz-landing-system : lando