ansible/hg-web: upgrade all virtualenvs to Python 3 and Mercurial 5.2 (Bug 1541166) r=smacleod
This commit upgrades all virtual environments on hgweb to Python 3.

The upgrade plan will involve deploying the repo state of v-c-t at
this commit to the hgweb mirrors in one AWS region, with the private
mirror rate for that region set to zero. Once the deploy has completed
successfully, we will set the rate to a low value (between 10-25% of
all CI traffic for that region) and verify there are no major bugs
with the deployment. We will then slowly increase the rate and ensure
the error rate for that region hasn't increased. Once we have increased
the rate to 100%, we will deploy this change to the remaining private
mirrors (set to 0 traffic, increasing directly to 100 once the deploy
completes successfully). Once our CI mirrors are fully upgraded to Py3,
we will conduct a rolling deployment to the public hg.mo interface in
MDC1, finally calling the upgrade complete.

We also upgrade `test-template-sync.t` to point to the new Mercurial
version being upgraded to, and change the hard-coded path to the template
folder to work with Python 3.

Differential Revision: https://phabricator.services.mozilla.com/D57348

--HG--
extra : moz-landing-system : lando