extensions: fix bug that clears extensions (Bug 1604074) r=sheehan
This change fixes a check that erased `cw.c['extensions']` if `b'extensions'`
is not present in `cw.c`. Main thing to note is that str and bytes can form
distinct keys in a dictionary, even though their string representation is the
same.

This change also fixes an issue that causes `mach` to crash if `evolve` fails
to enable, in which case the error message can not be interpreted correctly by
`ui`.

Differential Revision: https://phabricator.services.mozilla.com/D57650

--HG--
extra : moz-landing-system : lando