ansible/hg-ssh*: upgrade to Python 3 and Mercurial 5.2 (Bug 1541166) r=zeid,smacleod
This commit upgrades all virtual environments on hgssh to Python 3.

The upgrade plan for this commit is to request a tree closure window
from the sheriffs before conducting this upgrade. Once the trees are
closed, we will run through this deploy and wait until a successful
completion. After successful completion, we will make several test
pushes to user repositories to ensure no problems are encountered.
Once we have confidence that the upgrade has completed without a hitch,
we will re-open the trees and monitor performance of the system.

Differential Revision: https://phabricator.services.mozilla.com/D57349

--HG--
extra : moz-landing-system : lando